only fetching if not already fetched

+++ 3rdparty/fetch-mcps.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
-rm -rf MircryptionExtras*
-wget http://www.donationcoder.com/Software/Mouser/mircryption/downloads/MircryptionExtras.zip
-unzip MircryptionExtras.zip
+OLD_FOLDER=$(pwd)
+MY_FOLDER=$(dirname $0)
+
+cd $MY_FOLDER
+
+if [ ! -d MircryptionExtras ]
+then
+	rm -rf MircryptionExtras*
+	wget http://www.donationcoder.com/Software/Mouser/mircryption/downloads/MircryptionExtras.zip
+	unzip MircryptionExtras.zip
+fi
+
+cd $OLD_FOLDER
